Created Pages component and separated Spot form into two constants
