+++ frontend/src/components/Spot/CreateSpotForm.scss
@@ -9,7 +9,7 @@
         display: flex;
         flex-direction: column;
         align-items: center;
-        gap: 20px;
+        gap: 35px;
         
         .createSpotTitle {
             margin-top: 20px;
@@ -132,6 +132,16 @@
         }
     }
 
+    .nextPage {
+        height: 40px;
+        width: 150px;
+        border-radius: 5px;
+        margin-top: 40px;
+        background-color: #f8941983;
+        box-shadow: rgba(0, 0, 0, 0.15) 1.95px 1.95px 9px;
+        font-weight: 600;
+    }
+
     .image-preview {
         display: flex;
         gap: 20px;
@@ -163,6 +173,7 @@
     .rateType {
         display: flex;
         flex-direction: row;
+        margin-top: 30px;
     }
 
     input {
